requirements.txt
Appium-Python-Client==0.28
Faker==0.9.1
python-dateutil==2.7.3
selenium==3.14.1
six==1.11.0
text-unidecode==1.2
urllib3==1.23
